Refactored availability blob to description logic into supportCallService
