+++ app/teachingCallResponseReport/teachingCallResponseReportApp.js
@@ -14,6 +14,7 @@ import DwService from './../shared/services/DwService.js';
 import TermService from './../shared/services/TermService.js';
 import AuthService from './../shared/services/AuthService.js';
 import StringService from './../shared/services/StringService.js';
+import SupportCallService from './../shared/services/SupportCallService.js';
 
 // Dependencies
 var dependencies = [
@@ -57,6 +58,7 @@ const teachingCallResponseReportApp = angular.module("teachingCallResponseReport
 .service('StringService', StringService)
 .service('ApiService', ApiService)
 .service('TermService', TermService)
+.service('SupportCallService', SupportCallService)
 .constant('ActionTypes', {
 	INIT_STATE: "INIT_STATE"
 });
